Move KeystrokeSelectDialog to ui.components package
