+++ app/src/main/java/com/fake/soundremote/ui/components/KeystrokeSelectDialog.kt
@@ -1,4 +1,4 @@
-package com.fake.soundremote.ui.keystrokeselectdialog
+package com.fake.soundremote.ui.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -35,8 +35,6 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.fake.soundremote.R
 import com.fake.soundremote.ui.keystrokelist.KeystrokeListViewModel
 import com.fake.soundremote.ui.keystrokelist.KeystrokeUIState
-import com.fake.soundremote.ui.components.ListItemHeadline
-import com.fake.soundremote.ui.components.ListItemSupport
 
 @Composable
 internal fun KeystrokeSelectDialog(
